% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaedbns.R
\name{msaedbns}
\alias{msaedbns}
\title{EBLUPs under Multivariate Fay Herriot Model with Difference Benchmarking for non-sampled area}
\usage{
msaedbns(
  formula,
  vardir,
  weight,
  cluster,
  nonsample,
  samevar = FALSE,
  MAXITER = 100,
  PRECISION = 1e-04,
  data
)
}
\arguments{
\item{formula}{List of formula that describe the fitted model}

\item{vardir}{Sampling variances of direct estimations included in data frame as the vector with the name of sampling variances in order : \code{var1, cov12,.,cov1r,var2,cov23,.,cov2r,.,cov(r-1)(r),var(r)}}

\item{weight}{Known proportion of units in small areas, where \eqn{\sum_{d=1}^{D}}{sum from d=1 to D of} \eqn{W_{rd}}{Wrd} = 1 . \code{d = 1 ... D} is the number of small areas, and \code{r = 1 ... R} is the number of response variables}

\item{cluster}{cluster information}

\item{nonsample}{A column with logical values, \code{TRUE} if the area is non-sampled}

\item{samevar}{Whether the variances of the data are same or not. Logical input with default \code{FALSE}}

\item{MAXITER}{Maximum number of iteration in Fisher-scoring algorithm with default \code{100}}

\item{PRECISION}{Limit of Fisher-scoring convergence tolerance with default \code{1e-4}}

\item{data}{The data frame}
}
\value{
This function returns a list of the following objects:
   \item{MSAE_Eblup_sampled}{A dataframe with the values of the EBLUPs estimators for sampled areas}
   \item{MSAE_Eblup_all}{A dataframe with the values of the EBLUPs estimators for all areas}
   \item{MSE_Eblup_sampled}{A dataframe with the values of estimated mean square errors of EBLUPs estimators for sampled areas}
   \item{MSE_Eblup_all}{A dataframe with the values of estimated mean square errors of EBLUPs estimators for all areas}
   \item{randomEffect_sampled}{a dataframe with the values of the random effect estimators for sampled areas}
   \item{randomEffect_all}{a dataframe with the values of the random effect estimators for all areas}
   \item{Rmatrix_sampled}{a block diagonal matrix composed of sampling errors for sampled areas}
   \item{fit}{A list containing the following objects:}
     \itemize{
       \item method : The fitting method (this function is using "REML")
       \item convergence : The convergence result of Fisher-scoring algorithm (Logical Value)
       \item iterations : The number of Fisher-Scoring algorithm iterations
       \item estcoef : A dataframe with the estimated model coefficient, standard error,t statistics, p-values of the significance of each coefficient
       \item refvar : A dataframe with estimated random effect variances
       \item informationFisher : A matrix of information fisher from Fisher-scoring algorithm
     }

   \item{difference_benchmarking}{a list containing the following objects:}
     \itemize{
       \item Estimation_sampled : A dataframe with the values of benchmarked EBLUPs estimators for sampled areas
       \item Estimation_all : A dataframe with the values of benchmarked EBLUPs estimators for all areas
       \item Aggregation_sampled : The aggregation of benchmarked EBLUPs estimators, EBLUPs estimators and direct estimations for sampled areas
       \item Aggregation_all : The aggregation of benchmarked EBLUPs estimators, EBLUPs estimators and direct estimations for all areas
       \item MSE_DB_sampled : A dataframe with the values of estimated mean square errors of benchmarked EBLUPs estimators for sampled areas
       \item MSE_DB_all : A dataframe with the values of estimated mean square errors of benchmarked EBLUPs estimators for all areas
       \item g.4a : First component of g4 in difference benchmarking MSE estimation formula
       \item g.4b : Second component of g4 in difference benchmarking MSE estimation formula
     }
}
\description{
This function produces EBLUPs, MSE, and aggregation of Multivariate SAE with Difference Benchmarking for non-sampled area
}
\examples{
##load dataset
data(datamsaeDBns)

#Compute Fitted model for Y1, Y2, and Y3
#Y1 ~ X1 + X2
#Y2 ~ X1 + X2
#Y3 ~ X1 + X2

##Using parameter 'data'
formula = list(f1 = Y1~X1+X2,
               f2 = Y2~X1+X2,
               f3 = Y3~X1+X2)
vardir = c("v1","v12","v13","v2","v23","v3")
weight = c("w1","w2","w3")
cluster = c("clY1","clY2","clY3")
nonsample = "nonsample"
msaeDBns <- msaedbns(formula,vardir, weight,cluster, nonsample, data=datamsaeDBns)

}
